Moved connMan out of coco
