+++ connMan/chanConnManager.go
