+++ connMan/interface.go
