+++ insure/lifePolicy_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/dedis/crypto/config"
 	"github.com/dedis/crypto/random"
 
-	"github.com/dedis/prifi/coco/connMan"
+	"github.com/dedis/prifi/connMan"
 	"github.com/dedis/prifi/coconet"
 )
